Moved catch.hpp to lib/ folder.
